requirements.txt
# Core deep learning frameworks
torch>=2.8.0
torchvision>=0.23.0

# Computer vision and image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Scientific computing
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0

# Machine learning utilities
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0

# COCO evaluation tools
pycocotools>=2.0.6

# Progress bars
tqdm>=4.65.0

# Configuration files
PyYAML>=6.0.0

# Database (for inference pipeline)
# sqlite3 is included in Python standard library, no need to install

